server/pyproject.toml
[build-system]
requires = ["uv_build"]
build-backend = "uv_build"

[project]
name = "server"
version = "0.0.0"
requires-python = ">=3.12"
dependencies = [
    "ag-ui-langgraph",
    "fastapi[standard]",
    "geojson",
    "langchain-ollama",
    "pydantic-settings",
]

[project.scripts]
server = "server.main:main"
